Implemented the rewards distribution
